Fluid motion for movement, rotation, lift up; cone intake improvements
Cone intake rotates with or without cone 
Precise (slow) forward and rotational movement
Lift slows down slowly when moving up, reducing bouncing.

+++ Turning-PointV4/include/drive.h
@@ -13,8 +13,8 @@ public:
   float GetForwardAxis();
   float GetTurnAxis();
 
-  void SetLeftDrive(float speed);
-  void SetRightDrive(float speed);
+  void SetLeftDrive(int speed);
+  void SetRightDrive(int speed);
 
   void Update();
 };
